BUG: Include VTK libraries in Logic module
When SlicerOpenCV is packaged and installed, Slicer is showing various
errors related to the extension's Logic module. Most seem to be related to
VTK errors. Adding the VTK libraries to the list of target libraries explicitly,
following the example of the Chest Imaging Platform AirwayInspector extension
that was checked to work, this change results in a properly loaded
extension from the package file.

Issue #30