ENH: Add python wrapping to OpenCV
Turn on wrapping OpenCV in python.
This produces the cv2.so python package that can be imported.

Use the OpenCV CMake var -DPYTHON2_PACKAGES_PATH to specify that the
OpenCV python package needs to be installed in the third party library directory,
it will be found there by changes made to the AdditionalLauncherSettings.ini file
in Slicer svn 25305. To get packaged, it has to be in the Slicer third party
library install directory.

Adapted the opencv python example that computes the histogram of an input image to provide a self test.
Uploaded a copy of one of the Slicer Pathology sample slices to midas and download it for this test.

Don't install the ctest command line test on windows as it's not finding the python wrapped
library when the test is run from the build directory. It still runs from the Testing module
gui.

Issue #12